@@ -28,7 +28,7 @@
         "symfony/config": "~3.4|~4.0",
         "symfony/console": "~3.4|~4.0",
         "symfony/css-selector": "~3.4|~4.0",
-        "symfony/dependency-injection": "^4.1",
+        "symfony/dependency-injection": "^4.2",
         "symfony/dom-crawler": "~3.4|~4.0",
         "symfony/expression-language": "~3.4|~4.0",
         "symfony/finder": "~3.4|~4.0",
@@ -45,7 +45,7 @@
     },
     "conflict": {
         "symfony/config": "<3.4",
-        "symfony/dependency-injection": "<4.1",
+        "symfony/dependency-injection": "<4.2",
         "symfony/var-dumper": "<4.1",
         "twig/twig": "<1.34|<2.4,>=2"
     },
@@ -65,7 +65,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "4.1-dev"
+            "dev-master": "4.2-dev"
         }
     }
 }
