@@ -14,21 +14,40 @@ namespace Symfony\Component\Cache;
 use Psr\Cache\CacheItemInterface;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Cache\Exception\InvalidArgumentException;
+use Symfony\Component\Cache\Exception\LogicException;
 
 /**
  * @author Nicolas Grekas <p@tchwork.com>
  */
 final class CacheItem implements CacheItemInterface
 {
+    /**
+     * References the Unix timestamp stating when the item will expire.
+     */
+    const METADATA_EXPIRY = 'expiry';
+
+    /**
+     * References the time the item took to be created, in milliseconds.
+     */
+    const METADATA_CTIME = 'ctime';
+
+    /**
+     * References the list of tags that were assigned to the item, as string[].
+     */
+    const METADATA_TAGS = 'tags';
+
+    private const METADATA_EXPIRY_OFFSET = 1527506807;
+
     protected $key;
     protected $value;
     protected $isHit = false;
     protected $expiry;
     protected $defaultLifetime;
-    protected $tags = array();
-    protected $prevTags = array();
+    protected $metadata = array();
+    protected $newMetadata = array();
     protected $innerItem;
     protected $poolHash;
+    protected $isTaggable = false;
 
     /**
      * {@inheritdoc}
@@ -70,9 +89,9 @@ final class CacheItem implements CacheItemInterface
     public function expiresAt($expiration)
     {
         if (null === $expiration) {
-            $this->expiry = $this->defaultLifetime > 0 ? time() + $this->defaultLifetime : null;
+            $this->expiry = $this->defaultLifetime > 0 ? microtime(true) + $this->defaultLifetime : null;
         } elseif ($expiration instanceof \DateTimeInterface) {
-            $this->expiry = (int) $expiration->format('U');
+            $this->expiry = (float) $expiration->format('U.u');
         } else {
             throw new InvalidArgumentException(sprintf('Expiration date must implement DateTimeInterface or be null, "%s" given', is_object($expiration) ? get_class($expiration) : gettype($expiration)));
         }
@@ -86,11 +105,11 @@ final class CacheItem implements CacheItemInterface
     public function expiresAfter($time)
     {
         if (null === $time) {
-            $this->expiry = $this->defaultLifetime > 0 ? time() + $this->defaultLifetime : null;
+            $this->expiry = $this->defaultLifetime > 0 ? microtime(true) + $this->defaultLifetime : null;
         } elseif ($time instanceof \DateInterval) {
-            $this->expiry = (int) \DateTime::createFromFormat('U', time())->add($time)->format('U');
+            $this->expiry = microtime(true) + \DateTime::createFromFormat('U', 0)->add($time)->format('U.u');
         } elseif (\is_int($time)) {
-            $this->expiry = $time + time();
+            $this->expiry = $time + microtime(true);
         } else {
             throw new InvalidArgumentException(sprintf('Expiration date must be an integer, a DateInterval or null, "%s" given', is_object($time) ? get_class($time) : gettype($time)));
         }
@@ -109,14 +128,17 @@ final class CacheItem implements CacheItemInterface
      */
     public function tag($tags)
     {
-        if (!\is_array($tags)) {
+        if (!$this->isTaggable) {
+            throw new LogicException(sprintf('Cache item "%s" comes from a non tag-aware pool: you cannot tag it.', $this->key));
+        }
+        if (!\is_iterable($tags)) {
             $tags = array($tags);
         }
         foreach ($tags as $tag) {
             if (!\is_string($tag)) {
                 throw new InvalidArgumentException(sprintf('Cache tag must be string, "%s" given', is_object($tag) ? get_class($tag) : gettype($tag)));
             }
-            if (isset($this->tags[$tag])) {
+            if (isset($this->newMetadata[self::METADATA_TAGS][$tag])) {
                 continue;
             }
             if ('' === $tag) {
@@ -125,7 +147,7 @@ final class CacheItem implements CacheItemInterface
             if (false !== strpbrk($tag, '{}()/\@:')) {
                 throw new InvalidArgumentException(sprintf('Cache tag "%s" contains reserved characters {}()/\@:', $tag));
             }
-            $this->tags[$tag] = $tag;
+            $this->newMetadata[self::METADATA_TAGS][$tag] = $tag;
         }
 
         return $this;
@@ -135,10 +157,24 @@ final class CacheItem implements CacheItemInterface
      * Returns the list of tags bound to the value coming from the pool storage if any.
      *
      * @return array
+     *
+     * @deprecated since Symfony 4.2, use the "getMetadata()" method instead.
      */
     public function getPreviousTags()
     {
-        return $this->prevTags;
+        @trigger_error(sprintf('The "%s" method is deprecated since Symfony 4.2, use the "getMetadata()" method instead.', __METHOD__), E_USER_DEPRECATED);
+
+        return $this->metadata[self::METADATA_TAGS] ?? array();
+    }
+
+    /**
+     * Returns a list of metadata info that were saved alongside with the cached value.
+     *
+     * See public CacheItem::METADATA_* consts for keys potentially found in the returned array.
+     */
+    public function getMetadata(): array
+    {
+        return $this->metadata;
     }
 
     /**
