@@ -24,14 +24,11 @@ class PhpFilesCache extends AbstractCache implements PruneableInterface
      */
     public function __construct(string $namespace = '', int $defaultLifetime = 0, string $directory = null)
     {
-        if (!static::isSupported()) {
-            throw new CacheException('OPcache is not enabled');
-        }
+        self::$startTime = self::$startTime ?? $_SERVER['REQUEST_TIME'] ?? time();
         parent::__construct('', $defaultLifetime);
         $this->init($namespace, $directory);
 
         $e = new \Exception();
         $this->includeHandler = function () use ($e) { throw $e; };
-        $this->zendDetectUnicode = ini_get('zend.detect_unicode');
     }
 }
