@@ -15,7 +15,6 @@ use PHPUnit\Framework\TestCase;
 use Symfony\Component\DependencyInjection\Argument\IteratorArgument;
 use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Compiler\AnalyzeServiceReferencesPass;
-use Symfony\Component\DependencyInjection\Compiler\RepeatedPass;
 use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
@@ -207,7 +206,7 @@ class AnalyzeServiceReferencesPassTest extends TestCase
 
     protected function process(ContainerBuilder $container)
     {
-        $pass = new RepeatedPass(array(new AnalyzeServiceReferencesPass()));
+        $pass = new AnalyzeServiceReferencesPass();
         $pass->process($container);
 
         return $container->getCompiler()->getServiceReferenceGraph();
