@@ -243,7 +243,7 @@ class Symfony_DI_PhpDumper_Errored_Definition extends Container
      */
     protected function getFactoryServiceSimpleService()
     {
-        return $this->services['factory_service_simple'] = ($this->privates['factory_simple'] ?? $this->getFactorySimpleService())->getInstance();
+        return $this->services['factory_service_simple'] = $this->getFactorySimpleService()->getInstance();
     }
 
     /**
@@ -381,7 +381,7 @@ class Symfony_DI_PhpDumper_Errored_Definition extends Container
      */
     protected function getRuntimeErrorService()
     {
-        return $this->services['runtime_error'] = new \stdClass(($this->privates['errored_definition'] ?? $this->getErroredDefinitionService()));
+        return $this->services['runtime_error'] = new \stdClass($this->throw('Service "errored_definition" is broken.'));
     }
 
     /**
@@ -408,16 +408,6 @@ class Symfony_DI_PhpDumper_Errored_Definition extends Container
     }
 
     /**
-     * Gets the private 'errored_definition' shared service.
-     *
-     * @return \stdClass
-     */
-    protected function getErroredDefinitionService()
-    {
-        throw new RuntimeException('Service "errored_definition" is broken.');
-    }
-
-    /**
      * Gets the private 'factory_simple' shared service.
      *
      * @return \SimpleFactoryClass
@@ -428,7 +418,7 @@ class Symfony_DI_PhpDumper_Errored_Definition extends Container
     {
         @trigger_error('The "factory_simple" service is deprecated. You should stop using it, as it will soon be removed.', E_USER_DEPRECATED);
 
-        return $this->privates['factory_simple'] = new \SimpleFactoryClass('foo');
+        return new \SimpleFactoryClass('foo');
     }
 
     public function getParameter($name)
@@ -500,5 +490,10 @@ class Symfony_DI_PhpDumper_Errored_Definition extends Container
             'foo' => 'bar',
             'foo_bar' => 'foo_bar',
         );
+    }
+
+    protected function throw($message)
+    {
+        throw new RuntimeException($message);
     }
 }
