@@ -16,6 +16,7 @@ use Symfony\Component\HttpKernel\Event\GetResponseEvent;
 use Symfony\Component\HttpKernel\Event\FinishRequestEvent;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Symfony\Component\Security\Http\Firewall\LogoutListener;
 
 /**
  * Firewall uses a FirewallMap to register security listeners for the given
@@ -49,9 +50,14 @@ class Firewall implements EventSubscriberInterface
         // register listeners for this firewall
         $listeners = $this->map->getListeners($event->getRequest());
 
+        if (3 !== \count($listeners)) {
+            @trigger_error(sprintf('Not returning an array of 3 elements from %s::getListeners() is deprecated since Symfony 4.2, the 3rd element must be an instance of %s or null.', FirewallMapInterface::class, LogoutListener::class), E_USER_DEPRECATED);
+            $listeners[2] = null;
+        }
+
         $authenticationListeners = $listeners[0];
         $exceptionListener = $listeners[1];
-        $logoutListener = isset($listeners[2]) ? $listeners[2] : null;
+        $logoutListener = $listeners[2];
 
         if (null !== $exceptionListener) {
             $this->exceptionListeners[$event->getRequest()] = $exceptionListener;
